// src/api/services/views/types.ts
export interface IRemoteStaticDataWidget<T> extends TBaseRemoteWidget<T> {
    data_type: EWidgetData.Static;
    endpoint_url: null;
}

export interface IRemoteApiDataWidget<T> extends TBaseRemoteWidget<T> {
    data_type: EWidgetData.External | EWidgetData.Internal;
    endpoint_url: string;
}

export type IRemoteWidget<T = undefined> =
    | IRemoteApiDataWidget<T>
    | IRemoteStaticDataWidget<T>;

export type TRemoteUserViewWidgetSetting = {
    setting: {
        name: string;
        slug: string;
        setting_type: string;
    };
    tags: TRemoteTagReadOnly[];
    toggle_value: boolean;
};

export type TRemoteBaseUserViewWidget = {
    desktop_x: number;
    desktop_y: number;
    tablet_x: number;
    tablet_y: number;
    mobile_y: number;
};

export type TRemoteUserViewWidget<T = unknown> = TRemoteBaseUserViewWidget & {
    id: number;
    hash: string;
    widget: IRemoteWidget<T>;
    settings: TRemoteUserViewWidgetSetting[];
};

export type TRemoteBaseUserView = {
    name: string;
    description: string;
    is_favorite: boolean;
    is_system_view: boolean;
    slug?: string;
    sort_order: number;
    updated_at: string; // eg. 2022-06-17T10:49:43.659579Z
};

export type TRemoteUserView = TRemoteBaseUserView & {
    id: number;
    hash: string;
    icon: string | null;
    widgets: TRemoteUserViewWidget[];
    keywords: {
        id: number;
        name: string;
        tag: TRemoteTagReadOnly;
    }[];
    max_widgets: number;
};

export type TRemoteWidgetCategory = {
    id: number;
    name: string;
    slug: string;
    featured: boolean;
    sort_order: number;
};
// ...
export type TRemoteUserViewWidgetDraft = TRemoteBaseUserViewWidget & {
    hash?: string;
    widget: TRemoteWidgetMinValidate;
    settings: TRemoteUserViewWidgetSettingDraft[];
};

export type TRemoteUserViewDraft = TRemoteBaseUserView & {
    widgets: TRemoteUserViewWidgetDraft[];
    keywords: {
        id: number;
    }[];
};

/**
 * Query types
 */

export type TViewsRequest = void;
export type TViewsRawResponse = ReadonlyArray<TRemoteUserView>;
export type TViewsResponse = TViewsRawResponse;

export type TViewByIdRequest = { id: number };
export type TViewByIdRawResponse = Readonly<TRemoteUserView>;
export type TViewByIdResponse = TViewByIdRawResponse;

export type TViewByHashOrSlugRequest = { hash: string };
export type TViewByHashOrSlugResponse = TViewByIdResponse;

export type TSaveViewAsRequest = TRemoteUserViewDraft;
export type TSaveViewAsResponse = TRemoteUserView;

export type TSaveViewRequest = { id: number; body: TRemoteUserViewDraft };
export type TSaveViewResponse = TRemoteUserView;

export type TDeleteViewRequest = { id: number };
export type TDeleteViewResponse = null;

export type TWidgetsRequest = void;
export type TWidgetsRawResponse = ReadonlyArray<IRemoteWidget>;
export type TWidgetsResponse = TWidgetsRawResponse;

export type TWidgetByIdRequest = { id: number };
export type TWidgetByIdRawResponse = Readonly<IRemoteWidget>;
export type TWidgetByIdResponse = TWidgetByIdRawResponse;

export type TWidgetsCategoryRequest = {
    page?: number;
    limit?: number;
} | void;
export type TWidgetsCategoryResponse = TPagination & {
    results: ReadonlyArray<TRemoteWidgetCategory>;
};